requirements.txt
loguru==0.7.2
matplotlib==3.8.3
numpy==1.26.4
PyYAML==6.0.1
scikit_learn==1.4.1.post1
seaborn==0.13.2
tensorboardX==2.6.2.2
text2vec==1.3.0
torch==2.0.1
